#6 refactor localStorageMock
