+++ tests/game/GameRepository.test.ts
@@ -3,41 +3,11 @@ import { cleanup } from "@testing-library/react";
 import Game from '../../src/game/Game';
 import GameRepository from '../../src/game/GameRepository';
 import { JSDOM } from "jsdom";
+import localStorageMock from '../localStorageMock';
 
-// Set up JSDOM before running tests
 const jsdom = new JSDOM('<!doctype html><html><body></body></html>');
 (global as any).document = jsdom.window.document;
 (global as any).window = jsdom.window;
-
-const localStorageMock = (() => {
-   let store: { [key: string]: string } = {};
-
-   const localStorageInstance: Storage = {
-      getItem: function (key: string) {
-         return store[key] || null;
-      },
-      get: function () {
-         return store
-      },
-      setItem: function (key: string, value: string) {
-         store[key] = value;
-      },
-      removeItem: function (key: string) {
-         delete store[key];
-      },
-      clear: function () {
-         store = {};
-      },
-      key: function (index: number) {
-         return Object.keys(store)[index] || null;
-      },
-      get length() {
-         return Object.keys(store).length;
-      }
-   };
-
-   return localStorageInstance;
-})();
 
 global.localStorage = localStorageMock;
