+++ tests/game/useGame.test.tsx
@@ -5,40 +5,11 @@ import useGame from '../../src/game/useGame';
 import { JSDOM } from 'jsdom';
 import Game from '../../src/game/Game';
 import { ENHANCEMENTS, MISSIONS } from '../../src/game/constants';
+import localStorageMock from '../localStorageMock';
 
 const jsdom = new JSDOM('<!doctype html><html><body></body></html>');
 (global as any).document = jsdom.window.document;
 (global as any).window = jsdom.window;
-
-const localStorageMock = (() => {
-    let store: { [key: string]: string } = {};
-
-    const localStorageInstance: Storage = {
-        getItem: function (key: string) {
-            return store[key] || null;
-        },
-        get: function () {
-            return store
-        },
-        setItem: function (key: string, value: string) {
-            store[key] = value;
-        },
-        removeItem: function (key: string) {
-            delete store[key];
-        },
-        clear: function () {
-            store = {};
-        },
-        key: function (index: number) {
-            return Object.keys(store)[index] || null;
-        },
-        get length() {
-            return Object.keys(store).length;
-        }
-    };
-
-    return localStorageInstance;
-})();
 
 global.localStorage = localStorageMock;
